refactor(api): export a single app router
