--- packages/api/trpc-adapters/express.ts
@@ -1,1 +1,0 @@
-export * from "@trpc/server/adapters/express";
