--- packages/api/trpc-adapters/next.ts
@@ -1,1 +1,0 @@
-export * from "@trpc/server/adapters/next";
