+++ packages/api/tsconfig.json
@@ -4,11 +4,8 @@
     "src/**/*.ts",
     "scripts/**/*.ts",
     "test/**/*.ts",
-    "trpc-adapters/**/*.ts",
     "vitest.config.ts",
     "enums.ts",
-    "setup.ts",
-    "rest-api.ts",
-    "web.ts"
+    "setup.ts"
   ]
 }
